Fix file path in call to `clean_file`
The call to `i18n_tool transifex --pull` fails during the call to the
`clean_translated_locales` function because the path to the *.po files
is incorrect. In `dirname.files('*.po')`, the result already includes
the path relative to the `dirname`, so it is incorrect to perform an
additional `joinpath`.